I fucked up actually
Starting with directly 3d gameplay was a mistake. I deleted all of them, I'll first try to make a lobby UI, which is just 2d GUI. After I can manage to do that, i'll be working on the game. I'll create new branch for the lobby, u can work on actualy game if u want, leave networking to me.